requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
Flask-Cors==4.0.0
email-validator==2.0.0
flask-wtf==1.1.1